Move lighting code from GameScene
